% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast_crew_list.R
\name{cast_crew_list}
\alias{cast_crew_list}
\title{Retrieve TMDB cast and crew data for a given TMDB id vector.}
\usage{
cast_crew_list(id_col, api_key)
}
\arguments{
\item{id_col}{List of TMDB ids}

\item{api_key}{TMBD api key}
}
\value{
Dataframe containing all TMDB crew and cast data
}
\description{
Retrieve TMDB cast and crew data for a given TMDB id vector.
}
\examples{
\dontrun{
my_movie_ids <- data.frame(results.id = c(603, 604, 605))
cast_crew_list(my_movie_ids, api_key)
}
}
