% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_masterlist.R
\name{create_masterlist}
\alias{create_masterlist}
\title{Title}
\usage{
create_masterlist(wd, convention, api_key)
}
\arguments{
\item{wd}{Directory name containing your movie directories.}

\item{convention}{See examples.}

\item{api_key}{TMDB api key}
}
\value{
Extensive dataframe containing directory name, title, year and TMDB metadata for movie as well as cast.
}
\description{
Title
}
