% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details_list.R
\name{details_list}
\alias{details_list}
\title{Title}
\usage{
details_list(id_list, api_key)
}
\arguments{
\item{id_list}{List of TMDB ids}

\item{api_key}{TMDB api_key}
}
\value{
Dataframe containing all TMDB metadata for movies
}
\description{
Title
}
\examples{
\dontrun{
my_list <- c(603, 604, 605)
details_list(mylist, api_key)
}

}
