% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_batch_title_data.R
\name{extract_batch_title_data}
\alias{extract_batch_title_data}
\title{Title}
\usage{
extract_batch_title_data(movie_directory, convention)
}
\arguments{
\item{movie_directory}{The directory where movies are stored in.}

\item{convention}{Different naming conventions are supported. See extract_name.}
}
\value{
Dataframe containing directory name, movie year and title, optionally subtitles suffix (based on convention parameter).
}
\description{
Title
}
\examples{
\dontrun{
extract_batch_title_data("~/my_movie_folder/", 1)
}
}
