% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_names.R
\name{extract_names}
\alias{extract_names}
\title{Extract movie title and year from a vector}
\usage{
extract_names(name_vector, convention)
}
\arguments{
\item{name_vector}{Name(s) to extract title, year and potentially subtitle from.}

\item{convention}{See vignette.}
}
\value{
Dataframe containing directory name, movie year and title, optionally subtitles suffix (based on convention parameter).
}
\description{
Extract movie title and year from a vector
}
\examples{
extract_names("The Matrix (1999)", 3)

name_list <- c("The Matrix (1999)", "The Matrix Reloaded (2003)", "The Matrix Revolutions (2003)")
extract_names(name_list, 3)
}
