% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_title_data.R
\name{extract_title_data}
\alias{extract_title_data}
\title{Title}
\usage{
extract_title_data(directory_name, convention)
}
\arguments{
\item{directory_name}{Directory name to extract data from.}

\item{convention}{Different naming conventions are supported. See examples.}
}
\value{
Dataframe containing title, year and optionally subtitle abbreviation.
}
\description{
Title
}
\examples{
extract_title_data("1999 - The Matrix", 1)
extract_title_data("(1999) The Matrix", 2)
extract_title_data("The Matrix (1999)", 3)
extract_title_data("The Matrix (1999) (OmU)", 4)
}
