% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_credits.R
\name{get_credits}
\alias{get_credits}
\title{Title}
\usage{
get_credits(tmdb_id, api_key)
}
\arguments{
\item{tmdb_id}{TMDB id}

\item{api_key}{TMDB api key}
}
\value{
Dataframe containing TMDB credits (cast and crew).
}
\description{
Title
}
\examples{
\dontrun{
get_credits(603, api_key)
}
}
