% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_details.R
\name{get_details}
\alias{get_details}
\title{Title}
\usage{
get_details(tmdb_id, api_key)
}
\arguments{
\item{tmdb_id}{TMDB id}

\item{api_key}{TMDB api key}
}
\value{
Dataframe containing TMDB movie details
}
\description{
Title
}
\examples{
\dontrun{
get_details(603, api_key)
}
}
