% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genres.R
\name{get_genres}
\alias{get_genres}
\title{Title}
\usage{
get_genres(api_key)
}
\arguments{
\item{api_key}{TMDB api key}
}
\value{
Dataframe containing all TMDB genre ids and names.
}
\description{
Title
}
\examples{
\dontrun{
get_genres(api_key)
}
}
