% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{Title}
\usage{
get_id(title, year, api_key)
}
\arguments{
\item{title}{Movie title}

\item{year}{Release year}

\item{api_key}{TMDB api key}
}
\value{
TMDB id
}
\description{
Title
}
\examples{
\dontrun{
get_id("The Matrix", 1999, api_key)
}

}
