% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_list.R
\name{id_list}
\alias{id_list}
\title{Title}
\usage{
id_list(title_col, year_col, id_col, api_key)
}
\arguments{
\item{title_col}{list of titles to look up}

\item{year_col}{list of years}

\item{id_col}{list of identifiers}

\item{api_key}{TMDB api key}
}
\value{
Dataframe containing the movies directory name, title, year, and TMDB id
}
\description{
Title
}
